requirements.txt
streamlit
streamlit-drawable-canvas
numpy
Pillow
reportlab
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
